#98 - added traits for db adapter and security aware interface set up for test cases; replaced AbstractSecurityAwareTestCase
